Fix for GEOS-852, have JAI memory be a percentage of total memmory
git-svn-id: http://svn.codehaus.org/geoserver/trunk/geoserver@6018 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 05ae9908e27edd9a551341dd7b6aa8406fb3feda [formerly ad3a313f560b98d2ff02d88672b23b54d7344ffe]
Former-commit-id: 66112fc1ad0033151e34ec3b242523bc8da1688d

+++ main/src/main/java/org/vfny/geoserver/config/GlobalConfig.java
@@ -145,7 +145,7 @@ public class GlobalConfig {
 
     /** location on disk to log to **/
     private String logLocation = null;
-    private long jaiMemoryCapacity;
+    private double jaiMemoryCapacity;
     private double jaiMemoryThreshold;
     private int jaiTileThreads;
     private int jaiTilePriority;
@@ -619,11 +619,11 @@ public class GlobalConfig {
         this.loggingToFile = loggingToFile;
     }
 
-    public long getJaiMemoryCapacity() {
+    public double getJaiMemoryCapacity() {
         return jaiMemoryCapacity;
     }
 
-    public void setJaiMemoryCapacity(long jaiMemoryCapacity) {
+    public void setJaiMemoryCapacity(double jaiMemoryCapacity) {
         this.jaiMemoryCapacity = jaiMemoryCapacity;
     }
